% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyphil-package.R
\docType{package}
\name{nyphil-package}
\alias{nyphil}
\alias{nyphil-package}
\title{nyphil: New York Philharmonic Performance History Data}
\description{
Performance history data from all known concerts of the New York Philharmonic. Data were collected and made available by the New York Philharmonic.
}
\author{
\strong{Maintainer}: Aaron Butler \email{aaronbutlerphd@gmail.com} (\href{https://orcid.org/0000-0003-0433-3083}{ORCID})

}
\keyword{internal}
