% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyphil.R
\docType{data}
\name{nyphil}
\alias{nyphil}
\title{New York Philharmonic Performance History}
\format{
A tibble with 154,988 rows and 16 variables:
\describe{
\item{id}{a character string denoting the GUID; to view program: archives.nyphil.org/index.php/artifact/GUID/fullview}
\item{program_id}{a character string denoting the local NYP ID}
\item{orchestra}{a character string denoting the full orchestra name}
\item{season}{a character string denoting the season; defined as Sep 1 - Aug 31, displayed "1842-43"}
\item{event_type}{a character string denoting the name of event time; see: http://archives.nyphil.org/index.php/help-performancehistory}
\item{location}{a character string denoting the geographic location of concert; countries are identified by their current name. For example, even though the orchestra played in Czechoslovakia, it is now identified in the data as the Czech Republic}
\item{venue}{a character string denoting the name of hall, theater, or building where the concert took place}
\item{date}{a character string denoting the full ISO date used, but ignore TIME part (1842-12-07T05:00:00Z = Dec. 7, 1842)}
\item{time}{a character string denoting the actual time of concert, e.g. "8:00PM"}
\item{id_2}{a character string denoting the index number used to link soloist(s) and conductor(s) to a specific work on each of the concerts}
\item{composer_name}{a character string denoting the composer name}
\item{work_title}{a character string denoting the title of work; NYP short titles used}
\item{conductor_name}{a character string denoting the conductor name}
\item{soloist_name}{a character string denoting the soloist name (if multiple soloists on a single work, delimited by semicolon)}
\item{soloist_instrument}{a character string denoting the soloist instrument name}
\item{soloist_roles}{a character string denoting the soloist role: "S" means "Soloist"; "A" means "Assisting Artist" (if multiple soloists on a single work, delimited by semicolon)}
}
}
\source{
\url{https://archives.nyphil.org/performancehistory}
}
\usage{
nyphil
}
\description{
A tidy data version of the New York Philharmonic performance history data.
}
\keyword{datasets}
